fix(react-grid): update render component in correct lifecycle hook
