+++ packages/dx-react-grid-bootstrap3/src/plugins/virtual-table.jsx
@@ -17,7 +17,12 @@ export class VirtualTable extends React.PureComponent {
   constructor(props) {
     super(props);
 
-    this.layoutRenderComponent = createRenderComponent();
+    const { height, estimatedRowHeight } = props;
+    this.layoutRenderComponent =
+      createRenderComponent(VirtualTableLayout, { height, estimatedRowHeight });
+  }
+  componentWillReceiveProps({ height, estimatedRowHeight }) {
+    this.layoutRenderComponent.update({ height, estimatedRowHeight });
   }
   render() {
     const {
@@ -29,13 +34,7 @@ export class VirtualTable extends React.PureComponent {
 
     return (
       <TableBase
-        layoutComponent={this.layoutRenderComponent(props => (
-          <VirtualTableLayout
-            {...props}
-            height={height}
-            estimatedRowHeight={estimatedRowHeight}
-          />
-        ))}
+        layoutComponent={this.layoutRenderComponent.component}
         rowComponent={TableRow}
         cellComponent={TableCell}
         noDataRowComponent={TableRow}
